--- dbTemplete.js
@@ -1,8 +1,0 @@
-const db_config = {
-    host: '',
-    user: '',
-    password: '',
-    database: ''
-  };
-
-  module.exports = db_config;
